Fixed the captain issue. All methods of class Team have been tested. Added the new test to test_playground.py. Should implement binary insertion for the method add_new_kids and finish class Team with it.

+++ test_playground.py
@@ -52,10 +52,11 @@ class TestChildClass(unittest.TestCase):
 
 class TestTeamClass(unittest.TestCase):
 
+    # tests basic functionality of the class
     def test_constructor(self):
         for i in range(10, 1000, 5):
             Team.n_kids = i
-            Team.home_sender = i/5
+            Team.home_sender = int(i/5)
             if i % 2 == 0:
                 Team.Kid_problem_space = ProblemSpace1
             else:
